+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # Diccionario RAE Changelog
 
-## [Initial Version] - {PR_MERGE_DATE}
+## [Initial Version] - 2025-05-13
+
+- Added Command `Query RAE word`
+- Added Command `Query random RAE word`
+- Added Command `Query word of the day RAE word`